IOTDEV-89: Split index and show entry
